Make few tasks: * Remove atomic counter in shared state (we can use our state machine to determine when we need to remove it). atomic counter increment/decrement have a big cost in some scenarios * Decrease size of Future shared state to make it allocation cheaper * Some refactoring * Few fixes * Fix macOS and Windows CI * Add gcc-12 to CI

+++ src/fault/injector.cpp
@@ -33,7 +33,7 @@ bool Injector::NeedInject() noexcept {
 }
 
 void Injector::Reset() noexcept {
-  _count = GetRandNumber(sYieldFrequency);
+  _count = static_cast<uint32_t>(GetRandNumber(sYieldFrequency));
 }
 
 void Injector::SetFrequency(uint32_t freq) noexcept {
